adds interfaces for streaming calls, calls wait for Unary calls in Page
